chore(day-09): Fixed the issue/project/sprint deletion relative behavoior.

+++ Raman_Nema_Capstone_Issue_&_Sprint_Management_System/backend/app/services/project_service.py
@@ -7,6 +7,7 @@ from app.common.pagination import (
 )
 from app.models.project_model import ProjectModel
 from app.repositories.project_repository import ProjectRepository
+from app.repositories.sprint_repository import SprintRepository
 from app.repositories.user_repository import UserRepository
 from app.schemas.requests.project_member_request import (
     AssignMemberRequest,
@@ -21,6 +22,7 @@ from app.schemas.responses.project_response import (
 from app.exceptions.custom_exceptions import (
     ForbiddenException,
     ProjectAlreadyExistsException,
+    ProjectHasSprintsException,
     ProjectNotFoundException,
     UserNotFoundException,
     MemberAlreadyAssignedException,
@@ -218,6 +220,9 @@ class ProjectService:
 
         if not project:
             raise ProjectNotFoundException()
+
+        if SprintRepository.count_by_project(project_id) > 0:
+            raise ProjectHasSprintsException()
 
         ProjectRepository.delete_project(project_id)
